ui/services: Rearrange the services helpers
- we should put solution logic in services/k8s/solution while
  generic logic for metalk8s in services/k8s/core

Refs: #2278
