+++ ui/src/services/k8s/core.js
@@ -1,4 +1,5 @@
-import { coreV1 } from './api';
+import { coreV1, appsV1 } from './api';
+import jsonpatch from 'jsonpatch';
 
 export async function getNodes() {
   try {
@@ -71,6 +72,49 @@ export async function queryPodInNamespace(namespace, podLabel) {
       null,
       `app=${podLabel}`,
     );
+  } catch (error) {
+    return { error };
+  }
+}
+
+export async function createNamespacedConfigMap(name, namespace, restProps) {
+  const body = {
+    metadata: {
+      name,
+    },
+    restProps, //other props for V1ConfigMap: apiVersion, binaryData, data, kind
+  };
+  try {
+    return await coreV1.createNamespacedConfigMap(namespace, body);
+  } catch (error) {
+    return { error };
+  }
+}
+
+export async function patchNamespacedConfigMap(name, namespace, patch) {
+  // we don't have data prop in ConfigMap, in order to patch solution we need to initialize with a data object
+  const body = jsonpatch.apply_patch({ data: {} }, patch.jsonPatch);
+  try {
+    return await coreV1.patchNamespacedConfigMap(
+      name,
+      namespace,
+      body,
+      undefined,
+      undefined,
+      {
+        headers: {
+          'Content-Type': 'application/merge-patch+json',
+        },
+      },
+    );
+  } catch (error) {
+    return { error };
+  }
+}
+
+export async function getNamespacedDeployment(name, namespace) {
+  try {
+    return await appsV1.readNamespacedDeployment(name, namespace);
   } catch (error) {
     return { error };
   }
